[DDW-694] Fix Assets search
