+++ source/renderer/app/components/wallet/tokens/WalletTokensList.js
@@ -1,9 +1,8 @@
 // @flow
 import React from 'react';
 import { defineMessages, intlShape, injectIntl } from 'react-intl';
-import { omit, filter, escapeRegExp } from 'lodash';
 import { observer } from 'mobx-react';
-// import { onSearchAssetsDropdown } from '../../widgets/forms/AssetsDropdown';
+import { searchAssets } from '../../../utils/assets';
 
 import styles from './WalletTokens.scss';
 import Wallet from '../../../domains/Wallet';
@@ -29,32 +28,10 @@ type Props = {
   onAssetSettings: Function,
   onCopyAssetParam: Function,
   onOpenAssetSend: Function,
-  searchValue: string,
+  searchValue?: string,
   title: string,
   wallet: Wallet,
 };
-
-const onSearchAssetsDropdown = (_searchValue: string, assets: Array<any>) =>
-  filter(assets, (asset) => {
-    const searchValue = _searchValue.trim();
-    if (searchValue.length < 3) {
-      return true;
-    }
-    const { policyId, assetName, fingerprint, metadata } = asset;
-    const { name, ticker, description } = metadata || {};
-    const checkList = [
-      policyId,
-      assetName,
-      fingerprint,
-      metadata,
-      name,
-      ticker,
-      description,
-    ];
-    const regex = new RegExp(escapeRegExp(searchValue), 'i');
-    const result = checkList.some((item) => regex.test(item));
-    return result;
-  });
 
 const WalletTokensList = observer((props: Props) => {
   const {
@@ -66,13 +43,13 @@ const WalletTokensList = observer((props: Props) => {
     onAssetSettings,
     onCopyAssetParam,
     onOpenAssetSend,
-    searchValue,
+    searchValue = '',
     title,
     wallet,
   } = props;
   const isRestoreActive = wallet.isRestoring;
 
-  const filteredAssets = onSearchAssetsDropdown(searchValue, assets);
+  const filteredAssets = searchAssets(searchValue, assets);
 
   const noResults =
     !filteredAssets.length && searchValue && searchValue.trim().length >= 3;
