+++ source/renderer/app/components/widgets/forms/AssetsDropdown.js
@@ -1,8 +1,9 @@
 // @flow
 import React, { Component } from 'react';
-import { omit, filter, escapeRegExp } from 'lodash';
+import { omit } from 'lodash';
 import ItemsDropdown from './ItemsDropdown';
 import { formattedTokenWalletAmount } from '../../../utils/formatters';
+import { searchAssets } from '../../../utils/assets';
 import type { AssetToken } from '../../../api/assets/types';
 import Asset from '../../assets/Asset';
 import styles from './AssetsDropdown.scss';
@@ -20,34 +21,13 @@ type Props = {
   assets?: Array<$Shape<AssetToken>>,
 };
 
-export const onSearchAssetsDropdown = (
-  _searchValue: string,
-  options: Array<any>
-) => {
-  return filter(options, ({ asset }) => {
-    const searchValue = _searchValue.trim();
-    if (searchValue.length < 3) {
-      return true;
-    }
-    const { policyId, assetName, fingerprint, metadata } = asset;
-    const { name, ticker, description } = metadata || {};
-    const checkList = [
-      policyId,
-      assetName,
-      fingerprint,
-      metadata,
-      name,
-      ticker,
-      description,
-    ];
-    const regex = new RegExp(escapeRegExp(searchValue), 'i');
-    return checkList.some((item) => regex.test(item));
-  });
-};
-
 export default class AssetsDropdown extends Component<Props> {
   static defaultProps = {
-    onSearch: onSearchAssetsDropdown,
+    onSearch: (searchValue: string, assets: Array<any>) =>
+      searchAssets(
+        searchValue,
+        assets.map(({ asset }) => asset)
+      ),
   };
 
   render() {
